Add support for paginated results in list methods (#189)
This PR defines a new generic `Page` type, which represents a page of
results returned from methods like `replicate.models.list`. This PR also
updates those list methods to take a `cursor` argument. When `page.next`
or `page.previous` is not `None`, you can pass it to a `list` method to
get the next or previous page of results.

---------

Signed-off-by: Mattt Zmuda <mattt@replicate.com>